Do better error handling
Catch server errors and forward them up so it's shown
on the user's display. Also log them.

Better debug url with easy switching as well as version
information to help track down old versions.

+++ modules/slideshow.py
@@ -147,7 +147,8 @@ class slideshow:
           # We need to expunge any pending image now
           # so we get fresh data to show the user
           if self.imageCurrent:
-            os.remove(self.imageCurrent)
+            if os.path.exists(self.imageCurrent):
+              os.remove(self.imageCurrent)
             self.imageCurrent = None
             imageOnScreen = False
             self.display.clear()
